# Cache type information in Schema