… go.sum
… vendor/github.com/google/gofuzz/.travis.yml
… vendor/github.com/google/gofuzz/CONTRIBUTING.md
… vendor/github.com/google/gofuzz/LICENSE
… vendor/github.com/google/gofuzz/README.md
… vendor/github.com/google/gofuzz/doc.go
… vendor/github.com/google/gofuzz/fuzz.go
… vendor/github.com/google/gofuzz/go.mod
… vendor/modules.txt