When the schema is big, a lot of time is spent going through the list of each possible type to find the one it's looking for. No memory improvement, pure CPU/time:
```
benchmark                                   old ns/op     new ns/op     delta
BenchmarkOperations/Node/Update-12          576277        434979        -24.52%
BenchmarkOperations/Endpoints/Apply-12      2065245       1664846       -19.39%
BenchmarkOperations/Pod/Update-12           384166        312411        -18.68%
BenchmarkOperations/Node/Apply-12           764054        622976        -18.46%
BenchmarkOperations/Endpoints/Update-12     55470         46779         -15.67%
BenchmarkOperations/Pod/Apply-12            543377        478227        -11.99%
```