@apelisse 2019-10-02
New benchmark data:
```
BenchmarkOperations/Pod/Update-12                      384166        252547        -34.26%
BenchmarkOperations/Pod/Apply-12                       543377        390523        -28.13%
BenchmarkOperations/Node/Update-12                     576277        395116        -31.44%
BenchmarkOperations/Node/Apply-12                      764054        593929        -22.27%
BenchmarkOperations/Endpoints/Update-12                55470         31772         -42.72%
BenchmarkOperations/Endpoints/Apply-12                 2065245       1321254       -36.02%
BenchmarkFromUnstructured/Pod-12                       195262        125611        -35.67%
BenchmarkFromUnstructured/Node-12                      320953        272154        -15.20%
BenchmarkFromUnstructured/Endpoints-12                 8145833       6478752       -20.47%
```