@apelisse 2019-10-03
Updated, PTAL. Still not sure how to fix the test though :-/