@jennybuckley 2019-10-03
@apelisse its failing because the test uses quick.Check which does a reflect.DeepEqual?