@apelisse 2019-10-04
`quick.Check` doesn't want to deal with types that have unexported fields :-/. I guess I have to randomly generate values? I can probably use gofuzz instead